+++ RELEASE_NOTES.md
@@ -2,8 +2,9 @@
 
 ## Summary
 
-This is a patch release that updates the default host to use the FQDN instead of the IP address.
+This release adds support for pagination in the dispatch list request.
 
-## Bug Fixes
+## New Features
 
-* The default host was updated to use the FQDN instead of the IP address. This is required as the client validates the host name against the certificate.
+- Pagination support in the dispatch list request.
+